Updated Skimming and Dataset Managers. These are being used for current skim by Stefan and me. Will be merged with current crabconfig and kSkimming config before merging into master (when skim is completed succesfully)
